+++ src/main/java/com/lgcns/domain/visitorStats/batch/VisitorStatsScheduler.java
@@ -1,8 +1,7 @@
-package com.lgcns.infra.batch;
+package com.lgcns.domain.visitorStats.batch;
 
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
-import lombok.RequiredArgsConstructor;
 import org.springframework.batch.core.Job;
 import org.springframework.batch.core.JobParameters;
 import org.springframework.batch.core.JobParametersBuilder;
@@ -11,15 +10,23 @@ import org.springframework.batch.core.launch.JobLauncher;
 import org.springframework.batch.core.repository.JobExecutionAlreadyRunningException;
 import org.springframework.batch.core.repository.JobInstanceAlreadyCompleteException;
 import org.springframework.batch.core.repository.JobRestartException;
+import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Component;
 
 @Component
-@RequiredArgsConstructor
 public class VisitorStatsScheduler {
+
+    public static final String VISITOR_STATS_JOB = "visitorStatsJob";
 
     private final Job visitorStatsJob;
     private final JobLauncher jobLauncher;
+
+    public VisitorStatsScheduler(
+            @Qualifier(VISITOR_STATS_JOB) Job visitorStatsJob, JobLauncher jobLauncher) {
+        this.visitorStatsJob = visitorStatsJob;
+        this.jobLauncher = jobLauncher;
+    }
 
     @Scheduled(cron = "0 0 * * * *")
     public void createVisitorStats()
